requirements.txt
torch~=2.3.1+rocm6.0 # rocm version for AMD gpu
matplotlib~=3.9.0
numpy~=1.26.3
scikit-learn~=1.5.0
pandas~=2.2.2
requests~=2.32.3
torchvision~=0.18.1+rocm6.0 # rocm version for AMD gpu
torchmetrics~=1.4.0.post0
tqdm~=4.66.4
mlxtend~=0.23.1
pillow~=10.2.0